Auton and autoalign
